angypython/requirements.txt
gpsd-py3
requests
meshtastic
pubsub
websockets